Refactoring widgets a bit
